requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.22.0
Babel==2.4.0
billiard==3.3.0.23
celery==3.1.25
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
cloudinary==1.8.0
cryptography==2.0.2
defusedxml==0.5.0
dhl-shipping==2.0.4
Django==1.10
django-autoslug==1.9.3
django-autoslug-field==0.2.3
django-celery==3.2.1
django-ckeditor==5.3.0
django-colorfield==0.1.14
django-jenkins==0.110.0
django-jet==1.0.6
django-js-asset==0.1.1
django-multiselectfield==0.1.7
django-nested-admin==3.0.19
django-nested-inline==0.3.7
django-phonenumber-field==1.3.0
django-render-block==0.5
django-social-auth==0.7.28
django-templated-email==2.2.0
django-twilio==0.8.0
html5lib==1.0b10
httplib2==0.10.3
idna==2.5
jet==1.0
kombu==3.0.37
mock==2.0.0
ndg-httpsclient==0.4.2
oauth2==1.9.0.post1
oauthlib==2.0.2
olefile==0.44
pbr==3.1.1
phonenumberslite==8.7.1
Pillow==4.2.1
psycopg2==2.7.3
pusher==1.7.2
pyasn1==0.3.1
pycparser==2.18
PyJWT==1.5.0
pyOpenSSL==17.2.0
pyotp==2.2.6
PyPDF2==1.26.0
PySocks==1.6.7
python-dhl==1.0.0.dev14
python-monkey-business==1.0.0
python-openid==2.2.5
python-slugify==1.2.4
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2017.2
reportlab==3.4.0
requests==2.18.3
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
suds-jurko==0.6
twilio==6.5.0
Unidecode==0.4.21
urllib3==1.22
webencodings==0.5.1
whitenoise==3.3.0
xhtml2pdf==0.2b1
xmltodict==0.11.0
